requirements.txt
pika
influxdb-client
requests
python-dotenv